Move resolve_backend_index into the core ordering kernels
Share the user-to-backend index-resolution helper across backends by
defining it once in the core ordering-kernels module instead of the
Newton kernel module, so a future backend can reuse the same routing
with its identity fallback.
