+++ source/isaaclab/isaaclab/assets/articulation/ordering_kernels.py
@@ -46,7 +46,36 @@ __all__ = [
     "reorder_2d_user_to_backend",
     "reorder_3d_backend_to_user",
     "reorder_3d_user_to_backend",
+    "resolve_backend_index",
 ]
+
+
+@wp.func
+def resolve_backend_index(
+    user_id: wp.int32,
+    user_to_backend: wp.array(dtype=wp.int32),
+    has_ordering: bool,
+):
+    """Map a public (user-order) item index to its backend-order index.
+
+    Ordered write kernels that iterate the public joint or body axis and
+    scatter into backend-order buffers use this to centralize the
+    user-to-backend direction together with the identity fallback used when no
+    nonidentity ordering is active, so the routing logic is defined once
+    instead of re-implemented inline in each kernel.
+
+    Args:
+        user_id: Item index on the public (user-order) axis.
+        user_to_backend: Public-to-backend permutation for the axis. Ignored
+            when ``has_ordering`` is ``False`` (an identity map may be passed).
+        has_ordering: Whether a nonidentity ordering is active.
+
+    Returns:
+        The backend-order index for ``user_id``.
+    """
+    if has_ordering:
+        return user_to_backend[user_id]
+    return user_id
 
 
 @wp.kernel
